# docs/api/openapi-v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OmniFlow FastAPI 接口设计",
    "version": "1.0.0-draft",
    "description": "设计契约第一版，与Spec v1.0配套。尚未部署、不是从现有server.py或运行中的FastAPI生成；不得用文档成功代替真实接口验收。"
  },
...
  "servers": [
    {
      "url": "/api/v1",
      "description": "待实现的相对前缀；不表示现有域名已提供这些接口"
    }
  ],
...
  "security": [
    {
      "SessionCookie": []
    }
  ],
  "paths": {
...
    "/capabilities": {
      "get": {
        "operationId": "get_capabilities",
        "summary": "查询实际模型能力、可用性与安全上限",
        "tags": [
          "系统"
        ],
        "description": "查询实际模型能力、可用性与安全上限",
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "查询实际模型能力、可用性与安全上限",
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string",
                  "format": "uuid"
                },
                "description": "服务器生成的追踪ID"
              },
              "Cache-Control": {
                "schema": {
                  "type": "string"
                },
                "description": "受保护响应使用private, no-store"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Capabilities"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/HTTP400"
          },
          "401": {
            "$ref": "#/components/responses/HTTP401"
          },
          "403": {
            "$ref": "#/components/responses/HTTP403"
          },
          "429": {
            "$ref": "#/components/responses/HTTP429"
          },
          "500": {
            "$ref": "#/components/responses/HTTP500"
          },
          "503": {
            "$ref": "#/components/responses/HTTP503"
          }
        },
        "x-implementation-status": "design-only"
      }
    },
...
    "/conversations/{conversation_id}/reference-confirmations": {
      "post": {
        "operationId": "confirm_reference",
        "summary": "明确确认图片版本用于本对话首帧视频",
        "tags": [
          "作品"
        ],
        "description": "必须是自有可用图片版本。此动作是用户确认事实，不向MCP开放，模型不能自造确认。",
        "security": [
          {
            "SessionCookie": [],
            "CsrfHeader": []
          }
        ],
        "parameters": [
          {
            "name": "conversation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "responses": {
          "201": {
            "description": "明确确认图片版本用于本对话首帧视频",
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string",
                  "format": "uuid"
                },
                "description": "服务器生成的追踪ID"
              },
              "Cache-Control": {
                "schema": {
                  "type": "string"
                },
                "description": "受保护响应使用private, no-store"
              },
              "Idempotency-Replayed": {
                "schema": {
                  "type": "boolean"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReferenceConfirmation"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/HTTP400"
          },
          "401": {
            "$ref": "#/components/responses/HTTP401"
          },
          "403": {
            "$ref": "#/components/responses/HTTP403"
          },
          "404": {
            "$ref": "#/components/responses/HTTP404"
          },
          "409": {
            "$ref": "#/components/responses/HTTP409"
          },
          "410": {
            "$ref": "#/components/responses/HTTP410"
          },
          "422": {
            "$ref": "#/components/responses/HTTP422"
          },
          "429": {
            "$ref": "#/components/responses/HTTP429"
          },
          "500": {
            "$ref": "#/components/responses/HTTP500"
          },
          "503": {
            "$ref": "#/components/responses/HTTP503"
          }
        },
        "x-implementation-status": "design-only",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReferenceConfirmationCreate"
              }
            }
          }
        }
      }
    },
...
    "/uploads": {
      "post": {
        "operationId": "upload_image",
        "summary": "上传图片并建立不可变版本",
        "tags": [
          "作品"
        ],
        "description": "multipart/form-data；只接收可解码JPEG/PNG/WebP，限额见capabilities。幂等摘要使用文件字节及语义字段，不含multipart边界。",
        "security": [
          {
            "SessionCookie": [],
            "CsrfHeader": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "responses": {
          "201": {
            "description": "上传图片并建立不可变版本",
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string",
                  "format": "uuid"
                },
                "description": "服务器生成的追踪ID"
              },
              "Cache-Control": {
                "schema": {
                  "type": "string"
                },
                "description": "受保护响应使用private, no-store"
              },
              "Idempotency-Replayed": {
                "schema": {
                  "type": "boolean"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtifactCreated"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/HTTP400"
          },
          "401": {
            "$ref": "#/components/responses/HTTP401"
          },
          "403": {
            "$ref": "#/components/responses/HTTP403"
          },
          "409": {
            "$ref": "#/components/responses/HTTP409"
          },
          "410": {
            "$ref": "#/components/responses/HTTP410"
          },
          "413": {
            "$ref": "#/components/responses/HTTP413"
          },
          "415": {
            "$ref": "#/components/responses/HTTP415"
          },
          "422": {
            "$ref": "#/components/responses/HTTP422"
          },
          "429": {
            "$ref": "#/components/responses/HTTP429"
          },
          "500": {
            "$ref": "#/components/responses/HTTP500"
          },
          "503": {
            "$ref": "#/components/responses/HTTP503"
          },
          "507": {
            "$ref": "#/components/responses/HTTP507"
          }
        },
        "x-implementation-status": "design-only",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/UploadInput"
              }
            }
          }
        }
      }
    },
...
    "/artifacts": {
      "get": {
        "operationId": "list_artifacts",
        "summary": "列出自己的作品，不向模型默认注入其他会话作品",
        "tags": [
          "作品"
        ],
        "description": "列出自己的作品，不向模型默认注入其他会话作品",
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/PageCursor"
          },
          {
            "$ref": "#/components/parameters/PageLimit"
          },
          {
            "$ref": "#/components/parameters/ConversationFilter"
          },
          {
            "name": "kind",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "image",
                "video",
                "text"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "列出自己的作品，不向模型默认注入其他会话作品",
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string",
                  "format": "uuid"
                },
                "description": "服务器生成的追踪ID"
              },
              "Cache-Control": {
                "schema": {
                  "type": "string"
                },
                "description": "受保护响应使用private, no-store"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtifactPage"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/HTTP400"
          },
          "401": {
            "$ref": "#/components/responses/HTTP401"
          },
          "403": {
            "$ref": "#/components/responses/HTTP403"
          },
          "429": {
            "$ref": "#/components/responses/HTTP429"
          },
          "500": {
            "$ref": "#/components/responses/HTTP500"
          },
          "503": {
            "$ref": "#/components/responses/HTTP503"
          }
        },
        "x-implementation-status": "design-only"
      },
      "post": {
        "operationId": "create_text_artifact",
        "summary": "将文案另存为作品",
        "tags": [
          "作品"
        ],
        "description": "不调用模型，content存为受保护text/plain版本。",
        "security": [
          {
            "SessionCookie": [],
            "CsrfHeader": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "responses": {
          "201": {
            "description": "将文案另存为作品",
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string",
                  "format": "uuid"
                },
                "description": "服务器生成的追踪ID"
              },
              "Cache-Control": {
                "schema": {
                  "type": "string"
                },
                "description": "受保护响应使用private, no-store"
              },
              "Idempotency-Replayed": {
                "schema": {
                  "type": "boolean"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtifactCreated"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/HTTP400"
          },
          "401": {
            "$ref": "#/components/responses/HTTP401"
          },
          "403": {
            "$ref": "#/components/responses/HTTP403"
          },
          "409": {
            "$ref": "#/components/responses/HTTP409"
          },
          "410": {
            "$ref": "#/components/responses/HTTP410"
          },
          "422": {
            "$ref": "#/components/responses/HTTP422"
          },
          "429": {
            "$ref": "#/components/responses/HTTP429"
          },
          "500": {
            "$ref": "#/components/responses/HTTP500"
          },
          "503": {
            "$ref": "#/components/responses/HTTP503"
          },
          "507": {
            "$ref": "#/components/responses/HTTP507"
          }
        },
        "x-implementation-status": "design-only",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TextArtifactCreate"
              }
            }
          }
        }
      }
    },
    "/artifacts/{artifact_id}": {
      "get": {
        "operationId": "get_artifact",
        "summary": "读取作品及当前版本标识",
        "tags": [
          "作品"
        ],
        "description": "读取作品及当前版本标识",
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "parameters": [
          {
            "name": "artifact_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "读取作品及当前版本标识",
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string",
                  "format": "uuid"
                },
                "description": "服务器生成的追踪ID"
              },
              "Cache-Control": {
                "schema": {
                  "type": "string"
                },
                "description": "受保护响应使用private, no-store"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Artifact"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/HTTP400"
          },
          "401": {
            "$ref": "#/components/responses/HTTP401"
          },
          "403": {
            "$ref": "#/components/responses/HTTP403"
          },
          "404": {
            "$ref": "#/components/responses/HTTP404"
          },
          "429": {
            "$ref": "#/components/responses/HTTP429"
          },
          "500": {
            "$ref": "#/components/responses/HTTP500"
          },
          "503": {
            "$ref": "#/components/responses/HTTP503"
          }
        },
        "x-implementation-status": "design-only"
      },
      "delete": {
        "operationId": "delete_artifact",
        "summary": "撤销作品及所有版本访问并排队物理清理",
        "tags": [
          "作品"
        ],
        "description": "任务占用时409 RESOURCE_IN_USE；已生成的派生视频不连带删除。不承诺撤回已下载或已送达提供方的副本。",
        "security": [
          {
            "SessionCookie": [],
            "CsrfHeader": []
          }
        ],
        "parameters": [
          {
            "name": "artifact_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "撤销作品及所有版本访问并排队物理清理",
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string",
                  "format": "uuid"
                },
                "description": "服务器生成的追踪ID"
              },
              "Cache-Control": {
                "schema": {
                  "type": "string"
                },
                "description": "受保护响应使用private, no-store"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeletionReceipt"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/HTTP400"
          },
          "401": {
            "$ref": "#/components/responses/HTTP401"
          },
          "403": {
            "$ref": "#/components/responses/HTTP403"
          },
          "404": {
            "$ref": "#/components/responses/HTTP404"
          },
          "409": {
            "$ref": "#/components/responses/HTTP409"
          },
          "429": {
            "$ref": "#/components/responses/HTTP429"
          },
          "500": {
            "$ref": "#/components/responses/HTTP500"
          },
          "503": {
            "$ref": "#/components/responses/HTTP503"
          }
        },
        "x-implementation-status": "design-only"
      }
    },
    "/artifacts/{artifact_id}/versions": {
      "get": {
        "operationId": "list_artifact_versions",
        "summary": "分页读取作品历史版本",
        "tags": [
          "作品"
        ],
        "description": "分页读取作品历史版本",
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "parameters": [
          {
            "name": "artifact_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "$ref": "#/components/parameters/PageCursor"
          },
          {
            "$ref": "#/components/parameters/PageLimit"
          }
        ],
        "responses": {
          "200": {
            "description": "分页读取作品历史版本",
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string",
                  "format": "uuid"
                },
                "description": "服务器生成的追踪ID"
              },
              "Cache-Control": {
                "schema": {
                  "type": "string"
                },
                "description": "受保护响应使用private, no-store"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtifactVersionPage"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/HTTP400"
          },
          "401": {
            "$ref": "#/components/responses/HTTP401"
          },
          "403": {
            "$ref": "#/components/responses/HTTP403"
          },
          "404": {
            "$ref": "#/components/responses/HTTP404"
          },
          "429": {
            "$ref": "#/components/responses/HTTP429"
          },
          "500": {
            "$ref": "#/components/responses/HTTP500"
          },
          "503": {
            "$ref": "#/components/responses/HTTP503"
          }
        },
        "x-implementation-status": "design-only"
      }
    },
...
  },
  "components": {
...
    "parameters": {
...
      "PageCursor": {
        "name": "cursor",
        "in": "query",
        "required": false,
        "schema": {
          "type": "string",
          "minLength": 1,
          "maxLength": 512
        }
      },
...
    },
    "responses": {
      "HTTP400": {
        "description": "HTTP 400；稳定错误码与处理方式见接口文档",
        "headers": {
          "X-Request-ID": {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "服务器生成的追踪ID"
          },
          "Cache-Control": {
            "schema": {
              "type": "string"
            },
            "description": "受保护响应使用private, no-store"
          }
        },
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "HTTP429": {
        "description": "HTTP 429；稳定错误码与处理方式见接口文档",
        "headers": {
          "X-Request-ID": {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "服务器生成的追踪ID"
          },
          "Cache-Control": {
            "schema": {
              "type": "string"
            },
            "description": "受保护响应使用private, no-store"
          },
          "Retry-After": {
            "schema": {
              "type": "string",
              "pattern": "^[0-9]+$"
            },
            "description": "安全限流或队列背压；不是用户日配额"
          }
        },
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
...
      "HTTP403": {
        "description": "HTTP 403；稳定错误码与处理方式见接口文档",
        "headers": {
          "X-Request-ID": {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "服务器生成的追踪ID"
          },
          "Cache-Control": {
            "schema": {
              "type": "string"
            },
            "description": "受保护响应使用private, no-store"
          }
        },
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
...
      "HTTP401": {
        "description": "HTTP 401；稳定错误码与处理方式见接口文档",
        "headers": {
          "X-Request-ID": {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "服务器生成的追踪ID"
          },
          "Cache-Control": {
            "schema": {
              "type": "string"
            },
            "description": "受保护响应使用private, no-store"
          }
        },
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "SessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-omniflow_session",
        "description": "网站登录，不是Google/Agnes凭据"
      },
      "CsrfContextCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-omniflow_csrf",
        "description": "匿名表单上下文，不具有登录权限"
      },
      "CsrfHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token",
        "description": "绑定Cookie上下文，所有状态修改请求必须验证"
      },
      "MediaGrantToken": {
        "type": "apiKey",
        "in": "query",
        "name": "token",
        "description": "仅供应商取参考图片；后端生成短期高熵凭据，绑定grant_id、版本、任务和到期时间；不是网站或Google登录"
      }
    }
  },
...
}